+++ src/state/reducer.ts
@@ -6,7 +6,7 @@ import { abiApi } from './apis/abi/abiApi'
 import { fiatRampApi } from './apis/fiatRamps/fiatRamps'
 import { foxyApi } from './apis/foxy/foxyApi'
 import { limitOrderApi } from './apis/limit-orders/limitOrderApi'
-import { portals, portalsApi } from './apis/portals/portalsApi'
+import { portalsApi } from './apis/portals/portalsApi'
 import type { SnapshotState } from './apis/snapshot/snapshot'
 import { snapshot, snapshotApi } from './apis/snapshot/snapshot'
 import { swapperApi } from './apis/swapper/swapperApi'
@@ -170,7 +170,6 @@ export const apiSlices = {
   foxyApi,
   fiatRampApi,
   snapshotApi,
-  portals,
   portalsApi,
   opportunitiesApi,
   abiApi,
@@ -186,7 +185,6 @@ export const apiReducers = {
   [foxyApi.reducerPath]: foxyApi.reducer,
   [fiatRampApi.reducerPath]: fiatRampApi.reducer,
   [snapshotApi.reducerPath]: snapshotApi.reducer,
-  [portals.reducerPath]: portals.reducer,
   [portalsApi.reducerPath]: portalsApi.reducer,
   [opportunitiesApi.reducerPath]: opportunitiesApi.reducer,
   [abiApi.reducerPath]: abiApi.reducer,
